[fix] don't lose CHI cells without %pho cells
Used a wrong type of join (right instead of outer) when adding orphan pho's.
The fuzzy merge done by "merge_asof" can only do the left join so one more join was needed to add %phos that did not correspond to any of the CHIs (orphan %pho's).
I initially wanted to add orphans in an extra step so used the right join to get both orphan and parented %pho's but realized it was unnecessary and a single join would do.
However, it would have to be an outer join to keep the childless CHIs but I forgot to change the type.