handled pseudo-duplicates
It looked like there were pho cells corresponding to two CHI cells. This was only due to the matching tolerance of 0.5 s and CHI cell being closer than that to each other. It is rather obvious that only one of the CHI cell actually matches the pho cell - when you look at them manually.